server/pyproject.toml
[project]
name = "cinema-booking-backend"
version = "0.1.0"
description = "REST API сервиса бронирования билетов в кино"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.5",
    "email-validator>=2.3.0",
    "fastapi>=0.131.0",
    "greenlet>=3.3.2",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
]
